requirements.txt
cycler==0.10.0
dataclasses==0.6
future==0.18.2
joblib==0.17.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy~=1.18.5
opencv-python==4.4.0.46
pandas==1.1.5
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
torch==1.7.0+cu101
torchvision==0.8.1+cu101
typing-extensions==3.7.4.3

lmdb~=1.0.0
pytorchtools~=0.0.2